💩 refactor : useReactionData로 합치기
